Change "practice" section into "assignment" box
Normally, the exercises for each lesson are in a box under the "assignment" sub-header. This lesson uses "practice" as a sub-header instead, which migth throw off people that are only looking for assignments in each lesson. For this reason, and in general for consistency, I changed the sub-header title and wrapped the text in the usual "lesson-content__panel" div.